Changed removeTagsFromElement the relation really has to be deleted not set to "deleted" since the count happens after the delete it doesn't have to be decremented - even if we didn't really delete the relation the decrement wouldn't make any sense, since only the number of relations without the "deleted" flag was counted.
There is only one case, that needs a "deleted" flag for the relations table: That is, when a tagged element is set to "deleted".
In this case it must be possible to undelete the tagged element AND the relations to the tags later on. This is why we have to keep the tags in the DB regardless of their state within the relations table.