Added a caching mechanism to pmkshadowbox
This means that the original script files are merged into
a single script file. The cache file contents are depend on
the selected skin, language and adapter. Also the name of the cache
file is built from this three files.

Furthermore I have added a hook method to remove the cache files if
the clear cache button in the backend was clicked. This prevents
possible errors after a shadowbox update.

Note: All player scripts are included into the cache file always. They
are small and doesn't hurt anyone, but it prevents possible trouble and
too much cache files in the temporary directory. The temporary directory
can be found in typo3temp/. It's called simply pmkshadowbox.