Added caretaker_rsml extension with first implementation of "Remote Status Markup Language Tests"
Basically you can request an xml via url and get status informations from the file. The usecase is to connect caretaker to other external testscripts which are not part of a TYPO3 instance. 

RSML Example: 
<?xml version="1.0" encoding="UTF-8"?>
<rsml>
	<rsmlId>test</rsmlId>
	<rsmlVersion>1.2.3</rsmlVersion>
	<value>16</value>
	<status>0</status>
	<message>This is a message</message>
</rsml>

ATTENTION: 
You will be responsible to make shure that your testurl is not a security hole.