Feature #7770: checkCategoriesByRootline Option for checkCategoriesByRootline By default permissions for categories are builded via the complete rootline. This could cause an unexpected result. Therefore we added a ts option and disabled it by default.
checkCategoriesByRootline should be used, if most of the cats have no permissons and you want to touch as less permissons as possible.
The permisson check is done, by accumulation of all categories. Then an array of all usergroups is builded, which have access. If the current user is part of at least one group, than he has access.