[-FEATURE] Removed wkhtmltopdf binary
PDF ViewHelper now expects "wkhtmltopdf" to be in web server 
process' PATH - and if not, expects the full path to the binary to 
be specified as argument.

Download size of FED drastically decreased.

WARNING: This will break your code if you did NOT already install
wkhtmltopdf on your system.

See http://fedext.net/guides/shell-commands/support-for-wkhtmltopdf/

Fixes: #27211