[FEATURE] Possibility to delete answers and results
There is no possibility to delete answers and results when they are exported to a csv file. There is a demand for it, for instance when having the same survey for different groups, or for deleting test results.

This patch adds a new function to the module called "Delete results"

Fixes: #41461